refactor(benders): move OuterLoopBenders to outer_loop module
Move OuterLoopBenders from benders_mpi to the outer_loop module to
better reflect its logical ownership and decouple it from MPI-specific
concerns.

Key changes:
- Relocate OuterLoopBenders.h/.cpp from benders_mpi to outer_loop
- Update ICommunicationStrategy, SequentialCommunicationStrategy, and
  MpiCommunicationStrategy to support the decoupled interface required
  by OuterLoopBenders without a direct dependency on mpi::communicator
- Update benders_mpi/CMakeLists.txt to remove OuterLoopBenders sources
- Update outer_loop/CMakeLists.txt to include OuterLoopBenders sources
- Update BendersApp.cpp to use the new header location
- Update outer_loop_test.cpp to use the new header location
