+++ src/cpp/benders/outer_loop/CMakeLists.txt
@@ -21,9 +21,11 @@ endif ()
 add_library(outer_loop_lib)
 target_sources(outer_loop_lib PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/OuterLoop.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/OuterLoopBenders.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/OuterLoopBiLevel.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/outer_loop/IMasterUpdate.h
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/outer_loop/OuterLoop.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/outer_loop/OuterLoopBenders.h
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/outer_loop/OuterLoopBiLevel.h
 
 )
